optimize wavedata listening and fix interpolation for small fft sizes

+++ spectrum.h
@@ -49,7 +49,7 @@ struct spectrum_data_t {
     int num_channels;
     int num_samples;
     uint32_t channel_mask;
-    double *samples;
+    float *samples;
     double *spectrum;
     double *window;
     double *frequency;
